requirements.txt
python-bitget==1.1.2
python-dotenv
requests
